+++ libsrc/grabber/x11/X11Wrapper.cpp
@@ -4,14 +4,16 @@ X11Wrapper::X11Wrapper( int updateRate_Hz,
 						int pixelDecimation,
 						int cropLeft, int cropRight, int cropTop, int cropBottom)
 	: GrabberWrapper(GRABBERTYPE, &_grabber, updateRate_Hz)
-	  , _grabber(cropLeft, cropRight, cropTop, cropBottom)
-	  , _init(false)
+	, _grabber(cropLeft, cropRight, cropTop, cropBottom)
+	, _init(false)
 {
 	_grabber.setPixelDecimation(pixelDecimation);
 }
 
 X11Wrapper::X11Wrapper(const QJsonDocument& grabberConfig)
-	: GrabberWrapper(GRABBERTYPE, &_grabber)
+	: X11Wrapper(GrabberWrapper::DEFAULT_RATE_HZ,
+				 GrabberWrapper::DEFAULT_PIXELDECIMATION,
+				 0,0,0,0)
 {
 	this->handleSettingsUpdate(settings::SYSTEMCAPTURE, grabberConfig);
 }